# main-service/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Main service API
  version: '1.0'
servers:
  - description: Generated server url
    url: http://localhost:8080
...
paths:
  /admin/categories:
    post:
      description: 'Обратите внимание: имя категории должно быть уникальным'
      operationId: addCategory
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCategoryDto'
        description: данные добавляемой категории
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
          description: Категория добавлена
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Field: name. Error: must not be blank. Value: null'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление новой категории
      tags:
        - 'Admin: Категории'
  /admin/categories/{catId}:
    delete:
      description: 'Обратите внимание: с категорией не должно быть связано ни одного события.'
      operationId: deleteCategory
      parameters:
        - description: id категории
          in: path
          name: catId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Категория удалена
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=27 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Категория не найдена или недоступна
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: The category is not empty
                timestamp: '2023-01-21 16:56:19'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Существуют события, связанные с категорией
      summary: Удаление категории
      tags:
        - 'Admin: Категории'
    patch:
      description: 'Обратите внимание: имя категории должно быть уникальным'
      operationId: updateCategory
      parameters:
        - description: id категории
          in: path
          name: catId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        description: Данные категории для изменения
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryDto'
          description: Данные категории изменены
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=27 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Категория не найдена или недоступна
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Изменение категории
      tags:
        - 'Admin: Категории'
  /admin/compilations:
    post:
      operationId: saveCompilation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCompilationDto'
        description: данные новой подборки
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationDto'
          description: Подборка добавлена
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Field: title. Error: must not be blank. Value: null'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: 'could not execute statement; SQL [n/a]; constraint [uq_compilation_name]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление новой подборки (подборка может не содержать событий)
      tags:
        - 'Admin: Подборки событий'
  /admin/compilations/{compId}:
    delete:
      operationId: deleteCompilation
      parameters:
        - description: id подборки
          in: path
          name: compId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Подборка удалена
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Compilation with id=11 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Подборка не найдена или недоступна
      summary: Удаление подборки
      tags:
        - 'Admin: Подборки событий'
    patch:
      operationId: updateCompilation
      parameters:
        - description: id подборки
          in: path
          name: compId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCompilationRequest'
        description: данные для обновления подборки
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompilationDto'
          description: Подборка обновлена
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Category with id=27 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Подборка не найдена или недоступна
      summary: Обновить информацию о подборке
      tags:
        - 'Admin: Подборки событий'
...
  /admin/events/{eventId}:
    patch:
      description: |-
        Редактирование данных любого события администратором. Валидация данных не требуется.
        Обратите внимание:
         - дата начала изменяемого события должна быть не ранее чем за час от даты публикации. (Ожидается код ошибки 409)
        - событие можно публиковать, только если оно в состоянии ожидания публикации (Ожидается код ошибки 409)
        - событие можно отклонить, только если оно еще не опубликовано (Ожидается код ошибки 409)
      operationId: updateEvent_1
      parameters:
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventAdminRequest'
        description: Данные для изменения информации о событии
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие отредактировано
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=2 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: 'Cannot publish the event because it''s not in the right state: PUBLISHED'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Редактирование данных события и его статуса (отклонение/публикация).
      tags:
        - 'Admin: События'
  /admin/users:
    get:
      description: |-
        Возвращает информацию обо всех пользователях (учитываются параметры ограничения выборки), либо о конкретных (учитываются указанные идентификаторы)

        В случае, если по заданным фильтрам не найдено ни одного пользователя, возвращает пустой список
      operationId: getUsers
      parameters:
        - description: id пользователей
          in: query
          name: ids
          required: false
          schema:
            type: array
            items:
              type: integer
              format: int64
        - description: количество элементов, которые нужно пропустить для формирования текущего набора
          in: query
          name: from
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - description: количество элементов в наборе
          in: query
          name: size
          required: false
          schema:
            type: integer
            format: int32
            default: 10
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDto'
          description: Пользователи найдены
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
      summary: Получение информации о пользователях
      tags:
        - 'Admin: Пользователи'
    post:
      operationId: registerUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewUserRequest'
        description: Данные добавляемого пользователя
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
          description: Пользователь зарегистрирован
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Field: name. Error: must not be blank. Value: null'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated.
                message: 'could not execute statement; SQL [n/a]; constraint [uq_email]; nested exception is org.hibernate.exception.ConstraintViolationException: could not execute statement'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Нарушение целостности данных
      summary: Добавление нового пользователя
      tags:
        - 'Admin: Пользователи'
  /admin/users/{userId}:
    delete:
      operationId: deleteUser
      parameters:
        - description: id пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Пользователь удален
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: User with id=555 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Пользователь не найден или недоступен
      summary: Удаление пользователя
      tags:
        - 'Admin: Пользователи'
  /admin/events/{eventId}/comments/{commentId}:
    delete:
      description: 'Удаление комментария администратором'
      operationId: deleteCommentByAdmin
      parameters:
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
        - description: id комментария
          in: path
          name: commentId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Комментарий удален
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Comment with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Комментарий не найден
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: User is not allowed to delete this comment
                timestamp: '2023-01-21 16:56:19'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрещено удалять комментарии других пользователей
      summary: Удаление комментария администратором
      tags:
        - 'Admin: События'
...
  /users/{userId}/events/{eventId}:
    get:
      description: В случае, если события с заданным id не найдено, возвращает статус код 404
      operationId: getEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type long; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
      summary: Получение полной информации о событии добавленном текущим пользователем
      tags:
        - 'Private: События'
    patch:
      description: |
        Обратите внимание:
        - изменить можно только отмененные события или события в состоянии ожидания модерации (Ожидается код ошибки 409)
        - дата и время на которые намечено событие не может быть раньше, чем через два часа от текущего момента (Ожидается код ошибки 409)
      operationId: updateEvent
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id редактируемого события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateEventUserRequest'
        description: Новые данные события
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие обновлено
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: Event must not be published
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=283 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не найдено или недоступно
        '409':
          content:
            application/json:
              example:
                status: FORBIDDEN
                reason: For the requested operation the conditions are not met.
                message: Only pending or canceled events can be changed
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Событие не удовлетворяет правилам редактирования
      summary: Изменение события добавленного текущим пользователем
      tags:
        - 'Private: События'
...
  /users/{userId}/events/{eventId}/comments/{commentId}:
    get:
      description: 'Получение комментария пользователя'
      operationId: getComment
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
        - description: id комментария
          in: path
          name: commentId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullCommentDto'
          description: Комментарий изменен
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Comment with id=2 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос не найден или недоступен
      summary: Получение комментария по идентификатору
      tags:
        - 'Private: События'
    patch:
      description: 'Редактирование комментария пользователя'
      operationId: updateComment
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
        - description: id комментария
          in: path
          name: commentId
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CommentDto'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullCommentDto'
          description: Комментарий изменен
        '400':
          content:
            application/json:
              example:
                status: BAD_REQUEST
                reason: Incorrectly made request.
                message: 'Failed to convert value of type java.lang.String to required type int; nested exception is java.lang.NumberFormatException: For input string: ad'
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен некорректно
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Comment with id=2 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос не найден или недоступен
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: The category is not empty
                timestamp: '2023-01-21 16:56:19'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Пользователь не может удалять комментари других пользователей
      summary: Редактирование комментария
      tags:
        - 'Private: События'
    delete:
      description: 'Удаление комментария пользователем'
      operationId: deleteComment
      parameters:
        - description: id текущего пользователя
          in: path
          name: userId
          required: true
          schema:
            type: integer
            format: int64
        - description: id события
          in: path
          name: eventId
          required: true
          schema:
            type: integer
            format: int64
        - description: id комментария
          in: path
          name: commentId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Комментарий удален
        '404':
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Comment with id=13 was not found
                timestamp: '2022-09-07 09:10:50'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Комментарий не найден
        '409':
          content:
            application/json:
              example:
                status: CONFLICT
                reason: For the requested operation the conditions are not met.
                message: User is not allowed to delete this comment
                timestamp: '2023-01-21 16:56:19'
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрещено удалять комментарии других пользователей
      summary: Удаление комментария пользователем
      tags:
        - 'Private: События'
components:
  schemas:
...
    CategoryDto:
      required:
        - name
      type: object
      properties:
        id:
          type: integer
          description: Идентификатор категории
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          x-field-extra-annotation: "@javax.validation.constraints.NotBlank"
          description: Название категории
          example: Концерты
      description: Категория
    CompilationDto:
      required:
        - id
        - pinned
        - title
      type: object
      properties:
        events:
          uniqueItems: true
          type: array
          description: Список событий входящих в подборку
          example:
            - annotation: Эксклюзивность нашего шоу гарантирует привлечение максимальной зрительской аудитории
              category:
                id: 1
                name: Концерты
              confirmedRequests: 5
              eventDate: '2024-03-10 14:30:00'
              id: 1
              initiator:
                id: 3
                name: Фёдоров Матвей
              paid: true
              title: Знаменитое шоу 'Летающая кукуруза'
              views: 999
            - annotation: За почти три десятилетия группа 'Java Core' закрепились на сцене как группа, объединяющая поколения.
              category:
                id: 1
                name: Концерты
              confirmedRequests: 555
              eventDate: '2025-09-13 21:00:00'
              id: 1
              initiator:
                id: 3
                name: Паша Петров
              paid: true
              title: Концерт рок-группы 'Java Core'
              views: 991
          items:
            $ref: '#/components/schemas/EventShortDto'
        id:
          type: integer
          description: Идентификатор
          format: int64
          example: 1
        pinned:
          type: boolean
          description: Закреплена ли подборка на главной странице сайта
          example: true
        title:
          type: string
          description: Заголовок подборки
          example: Летние концерты
      description: Подборка событий
...
    NewCategoryDto:
      required:
        - name
      type: object
      properties:
        name:
          type: string
          x-field-extra-annotation: "@javax.validation.constraints.NotBlank"
          description: Название категории
          example: Концерты
      description: Данные для добавления новой категории
    NewCompilationDto:
      required:
        - title
      type: object
      properties:
        events:
          uniqueItems: true
          type: array
          description: Список идентификаторов событий входящих в подборку
          example:
            - 1
            - 2
            - 3
          items:
            type: integer
            description: Список идентификаторов событий входящих в подборку
            format: int64
        pinned:
          type: boolean
          description: Закреплена ли подборка на главной странице сайта
          example: false
          default: false
        title:
          type: string
          description: Заголовок подборки
          example: Летние концерты
      description: Подборка событий
...
    NewUserRequest:
      required:
        - email
        - name
      type: object
      properties:
        email:
          type: string
          description: Почтовый адрес
          example: ivan.petrov@practicummail.ru
        name:
          type: string
          description: Имя
          example: Иван Петров
      description: Данные нового пользователя
...